requirements.txt
azureml==0.2.7
beautifulsoup4==4.5.1
bs4==0.0.1
dj-database-url==0.4.1
Django==1.10.3
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
django-filter==1.0.1
djangorestframework==3.5.4
gunicorn==19.6.0
lxml==3.7.2
Markdown==2.6.8
numpy==1.11.1
oauthlib==1.1.2
pandas==0.18.1
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.11.0
requests-oauthlib==0.6.2
six==1.10.0
twitter==1.17.1
twython==3.4.0
virtualenv==15.0.2
whitenoise==3.2.1
psycopg2==2.6.2
python-decouple==3
